requirement.txt
fastapi==0.95.0
uvicorn==0.22.0
tweepy==4.12.1
google-api-python-client==2.70.0
nltk==3.8.1
python-dotenv==0.21.0
python-multipart==0.0.5
jinja2==3.1.2
